staging-imx: Fix enum_mbus_code return value
This operation is not implemented, but returning 0 means that it filled
the output argument with data for the requested index and there might be
more data to retrieve, causing userspace to continue calling the ioctl
over and over again.

Now this function returns -EINVAL, to indicate that there is no data to
be retrieved for the given index.