staging-imx: Hack to enable SRRGB10/RG10 format
This is very rough, it just adds the input and output format. The code
does not really know how to handle it, but accidentally seems to
configure the right handling (e.g. no YUV/RGB conversion). It also
relies on the CSI/sensor driver to force the input format, and does not
hardcoded the pixel order and bit depth. But it seems to work for now.