imx708: Crop output to 4096 pixels
This facilitates the IMX8MP ISI module, which can only process 4096
pixels wide images. If the sensor output is not cropped, the ISI will
throw away all but the leftmost 4096 pixels (this is unconfigurable).
By cropping in the sensor, we can throw away the outermost pixels, so
the cropping is centered.

This is now implemented by hardcoding different register values.
Relevant registers are 16-bit big-endian as follows:

0x0344-0x0345: X_ADD_STA_A
0x0346-0x0347: Y_ADD_STA_A
0x0348-0x0349: X_ADD_END_A
0x034A-0x034B: Y_ADD_END_A
0x034C-0x034D: x_output_size
0x034E-0x034F: x_output_size
0x040C-0x040D: x_output_size (dup)
0x040E-0x040F: x_output_size (dup)

Register info from: https://github.com/Hermann-SW/imx708_regs_annotated

It would be more elegant to either add a new cropped mode (this commit
removes support for the full resolution image) or, even better, control
the cropping from userspace (there is a selection API for this), but for
now this hardcoded cropping is probably sufficient (and a lot less
work), so this is tried first.