WIP: ar1335: Manually set exposure and gain
This driver has private IOCTLs for settings instead of supporting
standard v4l ctls, so setting gain and exposure is a bit cumbersome. So
instead, for initial testing, just set exposure and gain manuall.

The `ar1335_set_gain()` function accepts a gain multiplier (e.g. 10 is
10x gain, using analog where possible and then digital). The value
accepted by the `ar1335_set_exp()` function is a bit less clear, it is
written to the course integration time register as-is, so its
interpretation probably depends on other settings. This commit just sets
what looks to be the default setting.