3devo_wdt: Add module
This is a custom kernel module for the watchdog circuit on the FM23
actuatorboard v0.11. This circuit needs the watchdog to be pinged before
enabling regulators, but a dependency from a regulator onto a watchdog
module (especially not a *running* watchdog) is not something regulator
Linux drivers can model.

So this adds a driver that provides the watchdog driver and a single
virtual GPIO output which models the output of the watchdog circuit (and
input to regulators).