leds-pca963x: Add verbatim from mainline kernel
This takes the leds-pca963x driver from the toradex kernel, commit
bfdbfb2c85fb (arm64: configs: toradex: Enable TPM SPI driver, 2024-06-19)

This also adds a Makefile to compile it.

There is no need to rename any symbols or compatible strings, since this
is just a module that exposes no non-static symbols. In the toradex
kernel, it is also compiled as a module (not builtin), so if this is
included as a custom module in tcbuild.yaml, it will just replace the
default module.